destroyアクションの修正

+++ app/controllers/items_controller.rb
@@ -38,8 +38,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if item.destroy
-      redirect_to root_path
+    if current_user.id == item.user.id
+      if item.destroy
+        redirect_to root_path
+      else
+        render :show
+      end
     end
   end
